# contract/openapi/auth-service.yaml
openapi: 3.1.1
info:
  title: Auth Service
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /attributes:
    get:
      operationId: getAttributes
      parameters:
        - name: page
          in: query
          required: false
          schema:
            default: 0
            type: integer
        - name: size
          in: query
          required: false
          schema:
            default: 20
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            default: key ASC
            type: string
        - name: searchField
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributePage'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - attribute-controller
    post:
      operationId: addAttribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttributeData'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeDetail'
          description: Created
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - attribute-controller
  /attributes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteAttribute
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - attribute-controller
    get:
      operationId: getAttribute
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - attribute-controller
    put:
      operationId: setAttribute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttributeData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - attribute-controller
  /auth/change-email:
    post:
      operationId: changeEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmail'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/change-password:
    post:
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/change-user-attributes:
    post:
      operationId: changeUserAttributes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserAttributes'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/confirm:
    post:
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Confirmation'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/refresh:
    post:
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Refresh'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/resend-confirmation:
    post:
      operationId: resendConfirmation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendConfirmation'
        required: true
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
  /auth/reset-password:
    post:
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPassword'
        required: true
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
...
  /auth/user-detail:
    get:
      operationId: getUserDetail
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
          description: OK
        4XX:
          $ref: '#/components/responses/client-error'
        5XX:
          $ref: '#/components/responses/server-error'
      tags:
        - auth-controller
...
components:
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
  schemas:
    Page:
      type: object
      properties:
        totalElements:
          format: int64
          type: integer
        totalPages:
          format: int32
          type: integer
        page:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        sort:
          type: string
    AttributeDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        required:
          type: boolean
        hidden:
          type: boolean
    AttributePage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              items:
                $ref: '#/components/schemas/AttributeDetail'
              type: array
    ErrorCode:
      type: string
      enum:
        - UNKNOWN
        - NOT_FOUND
        - INVALID_BODY
        - INVALID_FIELD
        - INVALID_CAPTCHA
        - INVALID_CREDENTIALS
        - USER_NOT_ENABLED
        - USER_NOT_CONFIRMED
        - EMAIL_ALREADY_EXISTS
        - CANNOT_MANAGE_OWN_ACCOUNT
        - REQUIRED_ATTRIBUTE
    ErrorMessage:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        timestamp:
          format: date-time
          type: string
    AttributeData:
      type: object
      required:
        - key
        - required
        - hidden
      properties:
        key:
          type: string
        required:
          type: boolean
        hidden:
          type: boolean
    ChangeEmail:
      type: object
      required:
        - captchaText
        - captchaToken
        - email
        - password
      properties:
        email:
          format: email
          type: string
        password:
          type: string
        captchaText:
          type: string
        captchaToken:
          type: string
    AuthenticationResponse:
      type: object
      properties:
        refreshToken:
          type: string
        accessToken:
          type: string
    ChangePassword:
      type: object
      required:
        - captchaText
        - captchaToken
        - newPassword
        - oldPassword
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
        captchaText:
          type: string
        captchaToken:
          type: string
    AttributeValueData:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
    ChangeUserAttributes:
      type: object
      required:
        - attributes
        - captchaText
        - captchaToken
      properties:
        attributes:
          items:
            $ref: '#/components/schemas/AttributeValueData'
          type: array
        captchaText:
          type: string
        captchaToken:
          type: string
    Confirmation:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    Refresh:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    ResendConfirmation:
      type: object
      required:
        - captchaText
        - captchaToken
        - email
      properties:
        email:
          format: email
          type: string
        captchaText:
          type: string
        captchaToken:
          type: string
    ResetPassword:
      type: object
      required:
        - captchaText
        - captchaToken
        - email
      properties:
        email:
          format: email
          type: string
        captchaText:
          type: string
        captchaToken:
          type: string
...
    AttributeValueDetail:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
    AuthorityDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authority:
          type: string
    UserDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        confirmed:
          type: boolean
        enabled:
          type: boolean
        attributes:
          items:
            $ref: '#/components/schemas/AttributeValueDetail'
          type: array
        authorities:
          items:
            $ref: '#/components/schemas/AuthorityDetail'
          type: array
...
  responses:
    client-error:
      description: Client error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    server-error:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'